+++ smash.h
@@ -546,9 +546,9 @@ public:
 		//---------------------------------------------------------------
 		// (1) Cast the sphere into a 2 point square
 		RSphere* pSphere = &(pSmash->m_sphere->sphere);
-		long lR = pSphere->lRadius;
+      int32_t lR = pSphere->lRadius;
 
-		long	x1,x2,y1,y2;
+      int32_t	x1,x2,y1,y2;
 		x1 = x2 = pSphere->X;
 		y1 = y2 = pSphere->Z;
 		x1 -= lR;
@@ -606,12 +606,12 @@ public:
 		m_link4.m_pLast = pCurrent; // NOT YET INSTALLED THOUGH!
 		//-------------------------------------------------------------
 
-		//**************************************************
+      // **************************************************
 		// there is an error here! There could have been a 
 		// quadrant which did not CHANGE (so was not removed), 
 		// but is now REDUNDANT.  Think about combining these 
 		// sections!
-		//**************************************************
+      // **************************************************
 
 		// (4) Check for redundancies before installing:
 		// USE a NULL pointer to signify redundancy
